python-code/requirements.txt
azure-identity==1.10.0
azure-keyvault==4.2.0
requests==2.28.1